+++ R/AllGenerics.R
@@ -702,36 +702,63 @@ setGeneric("featureArea", function(object, ...) standardGeneric("featureArea"))
 #'
 #' @description
 #'
-#' Extract ion chromatograms for features in an [XcmsExperiment] or
-#' [XCMSnExp-class] object. The function returns for each feature the
-#' extracted ion chromatograms (along with all associated chromatographic
-#' peaks) in each sample. The chromatogram is extracted from the m/z - rt
-#' region that includes **all** chromatographic peaks of a feature. By default,
-#' this region is defined using the range of the chromatographic peaks' m/z
-#' and retention times (with `mzmin = min`, `mzmax = max`, `rtmin = min` and
-#' `rtmax = max`). For some features, and depending on the data, the m/z and
-#' rt range can thus be relatively large. The boundaries of the m/z - rt
-#' region can also be restricted by changing parameters `mzmin`, `mzmax`,
-#' `rtmin` and `rtmax` to a different functions, such as `median`.
+#' `featureChromatograms()` extracts ion chromatograms for features in an
+#' [XcmsExperiment] or [XCMSnExp-class] object. The function returns for each
+#' feature the extracted ion chromatograms in each sample. Based on the setting
+#' for parameter `return.type` the results are returned as a legacy
+#' [MSnbase::MChromatograms] or [XChromatograms()] object
+#' (`return.type = "MChromatograms"` or `return.type = "XChromatograms"`, the
+#' default) or as a new [Chromatograms::Chromatograms()] object
+#' (`return.type = "Chromatograms"`):
 #'
-#' By default only chromatographic peaks associated with a feature are
-#' included in the returned [XChromatograms] object. For `object` being an
-#' `XCMSnExp` object parameter `include` allows also to return all
-#' chromatographic peaks with their apex position within the selected
-#' region (`include = "apex_within"`) or any chromatographic peak overlapping
-#' the m/z and retention time range (`include = "any"`).
+#' - `return.type = "MChromatograms"`: the chromatograms are extracted from the
+#'   m/z - rt region that includes all chromatographic peaks for a feature. The
+#'   [featureArea()] function is used to define this region which is defined
+#'   using the range of the chromatographic peaks' m/z and retention times. The
+#'   `MChromatograms` organizes the EICs per feature, i.e., each row in the
+#'   returned `MChromatograms` corresponds to one feature with columns
+#'   containing EICs per sample.
+#'
+#' - `return.type = "XChromatograms"`: chromatographic data is defined as for
+#'   `return.type = "MChromatograms"`, but the returned EICs contain in addition
+#'   also the information on the individual chromatographic peaks as well as
+#'   the feature definitions. For `object` being an `XCMSnExp` object,
+#'   parameter `include` allows also to return all chromatographic peaks with
+#'   their apex position within the selected region (`include = "apex_within"`)
+#'   or any chromatographic peak overlapping the m/z and retention time range
+#'   (`include = "any"`).
+#'
+#' - `return.type = "Chromatograms"`: the EICs, defined through [featureArea()]
+#'   as described above, are returned as a [Chromatograms()] object which
+#'   organizes chromatograms sequentially in a list-like structure: first all
+#'   chromatograms for the first feature in all samples, then those from the
+#'   second feature and so on. See examples for details.
+#'   The feature ID is stored in the result object's `$feature_id` variable.
+#'   By default, the returned EICs represent the chromatograms of the
+#'   `featureArea()`, but with `featureArea = FALSE` each returned chromatogram
+#'   represents the EIC for the actual area of the individual chromatographic
+#'   peaks assigned to the feature (same as with [chromPeakChromatograms()]).
+#'   The IDs of the chromatographic peaks can then be accessed through
+#'   `$chrom_peak_id`.
 #'
 #' @note
 #'
-#' The EIC data of a feature is extracted from every sample using the same
-#' m/z - rt area. The EIC in a sample does thus not exactly represent the
-#' signal of the actually identified chromatographic peak in that sample.
-#' The [chromPeakChromatograms()] function would allow to extract the actual
-#' EIC of the chromatographic peak in a specific sample. See also examples
-#' below.
+#' Some of the functionality available for `XChromatograms` objects is not
+#' yet available for `Chromatograms`. Thus, while the newer
+#' `Chromatograms`-based infrastructure is more efficient and powerful, for
+#' some operations it is suggested to still use the legacy objects (such as
+#' simulating a correspondence analysis through [plotChromPeakDensity()]).
+#'
+#' By default the EIC data of a feature is extracted from every sample using
+#' the same m/z - rt area. Unless `featureArea = FALSE` is used, the EIC in a
+#' sample does thus not exactly represent the signal of the actually identified
+#' chromatographic peak in that sample.
 #'
 #' Parameters `include`, `filled`, `n` and `value` are only supported
 #' for `object` being an `XCMSnExp`.
+#'
+#' Parameter `featureArea` is only supported for
+#' `return.type = "Chromatograms"`.
 #'
 #' When extracting EICs from only the top `n` samples it can happen that one
 #' or more of the features specified with `features` are dropped because they
@@ -760,6 +787,10 @@ setGeneric("featureArea", function(object, ...) standardGeneric("featureArea"))
 #'
 #' @param expandRt `numeric(1)` to expand the retention time range for each
 #'     chromatographic peak by a constant value on each side.
+#'
+#' @param featureArea For `object` being a `XcmsExperiment` and
+#'     `return.type = FALSE`: return EICs representing data within the
+#'     identified chromatographic peaks.
 #'
 #' @param features `integer`, `character` or `logical` defining a subset of
 #'     features for which chromatograms should be returned. Can be the index
@@ -796,9 +827,9 @@ setGeneric("featureArea", function(object, ...) standardGeneric("featureArea"))
 #' @param progressbar `logical(1)` defining whether a progress bar is shown.
 #'
 #' @param return.type `character(1)` defining how the result should be
-#'     returned. At present only `return.type = "XChromatograms"` is
-#'     supported and the results are thus returned as an [XChromatograms()]
-#'     object.
+#'     returned. Supported are `return.type = "XChromatograms"` (the default),
+#'     `return.type = "MChromatograms"` and `return.type = "Chromatograms"`.
+#'     See function descriptions for details.
 #'
 #' @param rtmax `function` defining how the upper boundary of the rt region
 #'     from which the EIC is integrated should be defined. Defaults to
@@ -818,8 +849,11 @@ setGeneric("featureArea", function(object, ...) standardGeneric("featureArea"))
 #' @param ... optional arguments to be passed along to the [chromatogram()]
 #'     function.
 #'
-#' @return [XChromatograms()] object. In future, depending on parameter
-#'     `return.type`, the data might be returned as a different object.
+#' @return
+#'
+#' Depending on parameter `return.type`, a [MSnbase::MChromatograms],
+#' [XChromatograms] or [Chromatograms::Chromatograms()] object. See function
+#' description above for details.
 #'
 #' @name featureChromatograms
 #'
@@ -849,10 +883,51 @@ setGeneric("featureArea", function(object, ...) standardGeneric("featureArea"))
 #' ## Get the feature definitions
 #' featureDefinitions(xdata)
 #'
+#' #############################################################################
+#' ## Extracting the feature's EICs as an `Chromatograms` object
+#'
+#' ## Define the IDs for selected features from which to extract the data
+#' fids <- rownames(featureDefinitions(xdata))[1:3]
+#' chrs <- featureChromatograms(xdata, features = fids,
+#'     return.type = "Chromatograms")
+#' chrs
+#'
+#' ## The data is organized by feature and sample: first all EICs for the first
+#' ## feature in all samples, then those from the second etc
+#' chrs$feature_id
+#'
+#' ## The sample information is stored in the object's `$dataOrigin`
+#' basename(chrs$dataOrigin)
+#'
+#' ## Plot the data for the first feature
+#' library(Chromatograms)
+#'
+#' ## Separately
+#' plotChromatograms(chrs[chrs$feature_id == fids[1]])
+#'
+#' ## In a single plot
+#' plotChromatogramsOverlay(chrs[chrs$feature_id == fids[1]])
+#'
+#' ## Extract the EICs for the individual chromatographic peaks associated to
+#' ## each feature:
+#' chrs <- featureChromatograms(xdata, features = fids,
+#'     return.type = "Chromatograms", featureArea = FALSE)
+#' chrs
+#'
+#' ## In contrast to `featureArea = TRUE`, where one (and only one) EIC per
+#' ## sample is guaranteed, for `featureArea = FALSE` there can be more than
+#' ## one chromatogram per sample (or no chromatogram per sample), representing
+#' ## exactly the association between identified chromatographic peaks and
+#' ## features
+#' chrs$feature_id
+#' chrs$chrom_peak_id
+#'
+#' #############################################################################
+#' ## Using the legacy `MChromatogtrams`/`XChromatograms`
+#'
 #' ## Extract ion chromatograms for the first 3 features. Parameter
 #' ## `features` can be either the feature IDs or feature indices.
-#' chrs <- featureChromatograms(xdata,
-#'     features = rownames(featureDefinitions)[1:3])
+#' chrs <- featureChromatograms(xdata, features = fids)
 #'
 #' ## Plot the EIC for the first feature using different colors for each file.
 #' plot(chrs[1, ], col = c("red", "green", "blue"))